featureCrossReference now plural. Moved Ligand and LigandPart to main uniprot feature type file.
